Disable lyric editor when lyric events are edited
Whenever a command stack push or pop call is made, check if any lyric events were edited. If they were, close the lyric editor to prevent reference losses.

Note that direct SongEditCommand Invoke() and Revoke() calls are ignored, as they are used internally in the editor. As such, any tool which edits events without pushing those edits to the command stack should not be concurrently activated alongside the lyric editor.