Edit phrase content on left-click
Left-clicking on a phrase lets the user edit the contents of that phrase. That phrase and all following phrases will be picked up.

Note that newline characters are not restricted when editing this phrase; new phrases can be added here without restarting placement for the entire song, or entire phrases can be deleted by deleting their text content here.