Edit multiple phrases at once
When the user left-clicks on a phrase, they will be able to edit the lyrics of that phrase and the phrases following it. This serves two purposes: one, it ensures that no unplaced lyric is followed by a placed one, which is a state with undefined behavior currently; and two, it lets the user quickly remove multiple phrases, if needed.