Prevent parsing of user tags
For some inexplicable reason Unity does not have escape characters for angle brackets, so the next best thing is actual tags to break up its thinking. Simply replacing every "<" with "<<i></i>" does the trick at the expense of not being able to use italic text; this is okay, though, since I believe italics change text width, which could change line-ending positions as italics are toggled on and off.

Lyrics are still written exactly as-before, this just takes care of the rendering issues when interfering with color tags.