Change help button anchor point
Anchor to the top-left corner, rather than the bottom-right, so that resizing does not cause the button to leave the bounds of the editor panel. Not sure if you want to anchor directly to the target position instead FireFox, but I think this is the more logical solution if the background panel scales like I think it does.