Restore causality to the (CommandStack) universe
Differentiate between push and pop events to keep track of how many events remain in the stack, and exit the editor whenever an unrecognized ICommand is spotted. This guarantees that all recent CommandStack pushes will have come from the lyric editor, so they can just be popped off the end.